infra/requirements.txt
aws-cdk-lib>=2.243.0,<3.0.0
constructs>=10.5.0,<11.0.0
cdk-fck-nat>=1.6.0
